fix: duree_ans manchon préservé en float (0.5 → 6 mois, non 12)
- Frontend : supprime Math.round() dans FichePatientPage et OnboardingPage
- Backend : LPPRComponent.duree_ans Optional[int] → Optional[float]

+++ backend/server.py
@@ -83,7 +83,7 @@ class LPPRComponent(BaseModel):
     code: str
     nomenclature: str
     tarif: Optional[float] = None
-    duree_ans: Optional[int] = None
+    duree_ans: Optional[float] = None
     categorie: Optional[str] = None
     application: Optional[str] = None
     date_prescription: Optional[str] = None
